# openapi-users.yaml
openapi: 3.0.0
info:
  title: Healthcare Platform API
  version: 1.0.0
  description: |
    Complete API documentation for the healthcare platform's mobile and web applications.
    Supports member management, appointments, medical records, and e-commerce functionality.
servers:
  - url: https://api.healthcare-platform.com/v1
    description: Production server
  - url: https://staging-api.healthcare-platform.com/v1
    description: Staging server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Address:
      type: object
      properties:
        description:
          type: string
        pinCode:
          type: string
        region:
          type: string
        landmark:
          type: string
        state:
          type: string
        country:
          type: string
        location:
          type: object
          properties:
            latitude:
              type: number
            longitude:
              type: number
...
    MemberRegistration:
      type: object
      required:
        - name
        - phone
      properties:
        memberId:
          type: string
        AHIdentityCard:
          type: string
        isStudent:
          type: boolean
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        dob:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other]
        profilePic:
          type: string
          format: binary
        emergencyContact:
          type: object
          properties:
            name:
              type: string
            relation:
              type: string
            phone:
              type: string
        bloodGroup:
          type: string
        heightInFt:
          type: number
        weightInKg:
          type: number
        additionalInfo:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        employmentStatus:
          type: string
          enum: [employed, unemployed, self_employed, retired, student, homemaker]
        educationLevel:
          type: string
          enum: [primary, secondary, higher_secondary, graduate, post_graduate, doctorate]
        maritalStatus:
          type: string
          enum: [single, married, divorced, widowed]
        studentDetails:
          type: object
          properties:
            schoolId:
              type: string
            institution:
              type: string
            grade:
              type: string
            section:
              type: string
            guardians:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  relation:
                    type: string
                  phone:
                    type: string
            alternatePhone:
              type: number

    # Medical Related Schemas
...
paths:
...
  /members/profile:
    get:
      summary: Get member profile
      description: Get current member's profile details
      tags:
        - Members
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Member profile details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberRegistration'

    put:
      summary: Update member profile
      description: Update current member's profile details
      tags:
        - Members
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MemberRegistration'
      responses:
        '200':
          description: Profile updated successfully

  # Subprofile Endpoints
  /subprofiles:
    get:
      summary: Get all subprofiles
      description: Get list of all subprofiles for current member
      tags:
        - Subprofiles
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of subprofiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MemberRegistration'

    post:
      summary: Create new subprofile
      description: Create a new subprofile for family member
      tags:
        - Subprofiles
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MemberRegistration'
      responses:
        '201':
          description: Subprofile created successfully

  # Medical History Endpoints
...